+++ tests/testthat/helper-custom-fns.R
@@ -17,28 +17,78 @@ stand_up_custom_check_hub <- function(
     hub_path = system.file("testhubs/flusight", package = "hubValidations"),
     new_path = NULL,
     check_dir = testthat::test_path("testdata/src/R/"),
-    yaml_path = testthat::test_path("testdata/config/validations-src.yml")
-    ) {
+    yaml_path = testthat::test_path("testdata/config/validations-src.yml")) {
   if (is.null(new_path)) {
     return()
   }
   # make a copy of the hub
   fs::dir_copy(hub_path, new_path)
 
-  # create the script path 
+  # create the script path
   new_path <- fs::path(new_path, fs::path_file(hub_path))
   script_path <- fs::path(new_path, "src/validations/R")
   # NOTE: this creates the `src/validatons` directory because if the `R/` dir
   # exists, `dir_copy()` will place the copied directory _inside_ it, which is
-  # not what we want. 
+  # not what we want.
   fs::dir_create(fs::path_dir(script_path))
-  
+
   # Copy the directory of the existing files to the new script path
   fs::dir_copy(check_dir, script_path)
-  
+
   # Copy the provided yaml file to the config folder
   new_yaml <- fs::path(new_path, "hub-config", "validations.yml")
   fs::file_copy(yaml_path, new_yaml, overwrite = TRUE)
   # nolint end
   return(new_path)
+}
+
+# Create global target config_tasks for testing. Defaults to using task.json
+#  from clone of github.com/hubverse-org/example-complex-forecast-hub
+mock_global_target_config <- function(categorical = FALSE, config_tasks = NULL, hub_path = NULL) {
+  # Supplying config_tasks allows us to override any mocking that might be applied to
+  # read_config() in the tests.
+  if (is.null(config_tasks)) {
+    config_tasks <- read_config(hub_path)
+  }
+
+  if (categorical) {
+    # restrict to first round and model task 1 (target "wk flu hosp rate category")
+    config_tasks$rounds[[1]]$model_tasks <- config_tasks$rounds[[1]]$model_tasks[1]
+  } else {
+    # restrict to first round and model task 3 ("wk inc flu hosp" target)
+    config_tasks$rounds[[1]]$model_tasks <- config_tasks$rounds[[1]]$model_tasks[3]
+  }
+  # Assing NULL to target_keys
+  config_tasks <- purrr::assign_in(
+    config_tasks,
+    list(
+      "rounds", 1, "model_tasks", 1,
+      "target_metadata", 1, "target_keys"
+    ),
+    NULL
+  )
+  # Remove target task ID
+  config_tasks$rounds[[1]]$model_tasks[[1]]$task_ids[["target"]] <- NULL
+
+  config_tasks
+}
+
+mock_global_2_target_config <- function(config_tasks = NULL, hub_path = NULL) {
+  new_target <- "wk flu death rate category"
+  if (is.null(config_tasks)) {
+    config_tasks <- read_config(hub_path)
+  }
+  # Subset to single model task with distributional target
+  config_tasks$rounds[[1]]$model_tasks <- config_tasks$rounds[[1]]$model_tasks[1]
+
+  # Add target metadata for the new target
+  config_tasks$rounds[[1]]$model_tasks[[1]]$target_metadata[[2]] <-
+    config_tasks$rounds[[1]]$model_tasks[[1]]$target_metadata[[1]]
+  config_tasks$rounds[[1]]$model_tasks[[1]]$target_metadata[[2]]$target_id <-
+    new_target
+
+  # Add additional target to target task ID
+  config_tasks$rounds[[1]]$model_tasks[[1]]$task_ids[["target"]]$optional[2] <- new_target
+
+  config_tasks
 }